refactor and fixes
