+++ src/containers/Search/hooks/shared.ts
@@ -1,0 +1,26 @@
+export function merge(fromArr: SearchItem[], toArr: SearchItem[]) {
+  // maybe not good algorithm
+  const fromCids = fromArr.map((item) => item.cid);
+  const toCids = toArr.map((item) => item.cid);
+
+  const allCids = fromCids.filter((item) => toCids.includes(item));
+
+  const added: string[] = [];
+
+  return fromArr.concat(toArr).reduce<SearchItem[]>((acc, item) => {
+    if (added.includes(item.cid)) {
+      return acc;
+    }
+
+    if (allCids.includes(item.cid)) {
+      added.push(item.cid);
+
+      return acc.concat({
+        ...item,
+        type: 'all',
+      });
+    }
+
+    return acc.concat(item);
+  }, []);
+}
